r/virtual_machine: Storage DRS recommendations for VM creation
This adds functional support for storage DRS recommendations on VM
creation.

We are not using the storage DRS API for actual VM creation for this
process - rather we adopt an approach that is similar to govc currently,
where the recommendations are requested for an existing configspec and
and then directly re-applied to the config spec, with the VM then being
created off of that.

Some minor adjustments needed to be made to the virtual device code to
account for missing datastores during the creation process, but
otherwise this process was pretty non-invasive.

This process of actually side-stepping applying recommendations through
the API is probably what we will use for most of the storage DRS support
as there are a couple of situations where we would not be able to apply
recommendations atomically, like in VM cloning where a disk is being
grown post-clone. Conversely, there may be situations where we will
actually need to use the API, such as during reconfiguration operations,
if migrating a disk to a new datastore will actually allow a
reconfiguration operation to complete. This latter use case is still
TBD.